4_batch_processing/1_model_training/requirements.txt
# PyTorch CPU-only (lighter than CUDA version)
torch
torchvision

# Core dependencies
ultralytics
minio
Pillow
PyYAML
pyspark==3.4.1

